chore: Move insights-client management to own package
* Card ID: CCT-1589

To improve code separation, move the insights-client related methods
to their own package.
